requirements.txt
streamlit>=1.45.0
requests>=2.31.0
openai>=1.78.0
python-dotenv>=1.0.1
